% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\docType{data}
\name{adjustment_models}
\alias{adjustment_models}
\title{A list of adjustment model formulas}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43 rows and 3 columns.
}
\usage{
adjustment_models
}
\description{
A list of adjustment model formulas
}
\keyword{datasets}
