% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\docType{data}
\name{adjustment_models_diet_1}
\alias{adjustment_models_diet_1}
\title{A list of adjustment model formulas for dietary exposures (day 1)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 3 columns.
}
\usage{
adjustment_models_diet_1
}
\description{
A list of adjustment model formulas for dietary exposures (day 1)
}
\keyword{datasets}
