% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\docType{data}
\name{adjustment_models_ucr}
\alias{adjustment_models_ucr}
\title{A list of adjustment model formulas for urine exposures}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 3 columns.
}
\usage{
adjustment_models_ucr
}
\description{
A list of adjustment model formulas for urine exposures
}
\keyword{datasets}
