% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{adjustment_scenario_for_variable}
\alias{adjustment_scenario_for_variable}
\title{Determine Adjustment Scenario for a Given Exposure and Phenotype}
\usage{
adjustment_scenario_for_variable(evarname, pvarname)
}
\arguments{
\item{evarname}{A character string specifying the name of the exposure variable.}

\item{pvarname}{A character string specifying the name of the phenotype variable.}
}
\value{
The function returns one of the predefined adjustment model sets based on the variable name patterns:
\itemize{
\item \code{adjustment_models_diet_x}: If the exposure variable starts with \code{"DRX"}, indicating diet-related data adjusted by total calories.
\item \code{adjustment_models_diet_1}: If the exposure variable starts with \code{"DR1"}, indicating first-day diet recall data adjusted by total calories.
\item \code{adjustment_models_diet_2}: If the exposure variable starts with \code{"DR2"}, indicating second-day diet recall data adjusted by total calories.
\item \code{adjustment_models_ucr}: If the exposure variable starts with \code{"UR"} and the phenotype variable does not start with \code{"UR"}, indicating urine data adjusted by creatinine.
\item \code{adjustment_models}: A default set of adjustment models if none of the above conditions are met.
}
}
\description{
This function determines the appropriate adjustment scenario based on the names of the exposure and phenotype variables. It returns a predefined set of adjustment models specific to the type of variable provided.
}
\details{
The function uses the first few characters of the exposure and phenotype variable names to determine the appropriate adjustment scenario. It logs the chosen adjustment type using \code{log_info}.
}
\examples{
\dontrun{
evarname <- "DR1TKCAL"
pvarname <- "BMXBMI"
adjustment_model <- adjustment_scenario_for_variable(evarname, pvarname)
print(adjustment_model)
}
}
