% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\docType{data}
\name{adjustmentVariables}
\alias{adjustmentVariables}
\title{A list of adjustmentVariables that match those in the demographic (e.g., demo tables)}
\format{
An object of class \code{character} of length 12.
}
\usage{
adjustmentVariables
}
\description{
A list of adjustmentVariables that match those in the demographic (e.g., demo tables)
}
\keyword{datasets}
