% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petable.R
\name{check_e_data_type}
\alias{check_e_data_type}
\title{Check Exposure Data Type}
\usage{
check_e_data_type(varname, con)
}
\arguments{
\item{varname}{A character string specifying the name of the variable to be checked.}

\item{con}{A database connection object used to query the database for variable levels.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{vartype}: A character string indicating the type of the variable. It can be one of \code{"continuous"}, \code{"continuous-rank"}, or \code{"categorical"}.
\item \code{varlevels}: A vector of variable levels if the variable is categorical or continuous-rank, otherwise \code{NULL}.
}
}
\description{
This function checks the data type of a given exposure variable based on its name and associated levels stored in a database. It returns information about whether the variable is continuous, continuous-rank, or categorical.
}
\details{
The function determines the variable type based on the following rules:
\itemize{
\item If the variable name ends with \code{"CNT"}, it is treated as \code{"continuous-rank"}.
\item If the variable name starts with \code{"PAQ"}, it is treated as \code{"continuous"}.
\item If the variable has a single unique level, it is treated as \code{"continuous"}.
\item If the variable levels are non-integer values, it is treated as \code{"continuous-rank"}.
\item If the variable levels are all integer values, it is treated as \code{"categorical"}.
}
}
\examples{
\dontrun{
con <- DBI::dbConnect(RSQLite::SQLite(), dbname = "exposome.db")
result <- check_e_data_type("LBXCOT", con)
print(result$vartype)
}
}
