% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantpe.R
\name{connect_pewas_data}
\alias{connect_pewas_data}
\title{Connect to the PEWAS NHANES database}
\usage{
connect_pewas_data(path_to_data = NULL)
}
\arguments{
\item{path_to_data}{The file path to the participant-level NHANES data. Each table in the database corresponds to a table in the NHANES.}
}
\value{
A DBI connection object to the NHANES database.
}
\description{
This function establishes a connection to the PEWAS NHANES database .
}
\examples{
\dontrun{
con <- connect_pewas_data("/path/to/database.sqlite")
}
}
