% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantpe.R
\name{connect_pewas_nhanes_database}
\alias{connect_pewas_nhanes_database}
\title{Connect to the PEWAS NHANES database}
\usage{
connect_pewas_nhanes_database(path_to_database)
}
\arguments{
\item{path_to_database}{The file path to the SQLite database}
}
\value{
A DBI connection object to the SQLite database.
}
\description{
This function establishes a connection to the PEWAS NHANES database using SQLite.
}
\examples{
\dontrun{
con <- connect_pewas_nhanes_database("/path/to/database.sqlite")
}
}
