% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantpe.R
\name{demographic_breakdown}
\alias{demographic_breakdown}
\title{Provide a demographic breakdown for a survey design object}
\usage{
demographic_breakdown(svy_dsn)
}
\arguments{
\item{svy_dsn}{A survey design object derived from the nhanespewas::get_*_tables() functions}
}
\value{
A tibble containing the mean for each demographic variable in the survey design object.
}
\description{
This function calculates the mean for several hard-coded demographic variables in the survey design object.
The demographic variables are: RIDAGEYR, RIAGENDR, INDFMPIR, ETHNICITY_MEXICAN, ETHNICITY_OTHERHISPANIC,
ETHNICITY_OTHER, ETHNICITY_NONHISPANICBLACK, ETHNICITY_NONHISPANICWHITE, EDUCATION_LESS9, EDUCATION_9_11,
EDUCATION_HSGRAD, EDUCATION_AA, and EDUCATION_COLLEGEGRAD.
}
