% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantpe.R
\name{disconnect_database}
\alias{disconnect_database}
\title{Disconnect from a database}
\usage{
disconnect_database(dbConn)
}
\arguments{
\item{dbConn}{A DBI connection object.}
}
\value{
Invisible NULL. The connection will be completely severed.
}
\description{
This function disconnects from a database given a DBI connection object.
}
\examples{
\dontrun{
disconnect_database(con)
}
}
