% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petable.R
\name{disconnect_pewas_data}
\alias{disconnect_pewas_data}
\title{Disconnect from a NHANES database}
\usage{
disconnect_pewas_data(dbConn)
}
\arguments{
\item{dbConn}{A NHANES DBI connection object.}
}
\value{
Invisible NULL. The connection will be completely severed.
}
\description{
This function disconnects from a database given a DBI connection object.
}
\examples{
\dontrun{
disconnect_pewas_data(dbConn)
}
}
