% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petable.R
\name{figure_out_multiyear_weight}
\alias{figure_out_multiyear_weight}
\title{Figure Out Multiyear Weight}
\usage{
figure_out_multiyear_weight(get_tables_as_list_obj)
}
\arguments{
\item{get_tables_obj}{A list object that contains the data tables and relevant metadata. This object should include:
\itemize{
\item \code{series}: A data frame that contains the series information.
\item \code{weight_name}: The name of the original weight variable.
\item \code{merged_tab}: The merged data table that contains the survey data.
}}
}
\value{
The updated \code{get_tables_obj} list with recalculated weights.
}
\description{
This function calculates appropriate weights for multiyear survey data, specifically handling special cases where both series A and B are present. It updates the \code{weight_name} and adjusts weights in the merged table accordingly.
}
\details{
The function performs the following steps:
\itemize{
\item Calls \code{figure_out_weight} to initialize weights.
\item Checks if both series A and B are present. If so, it logs the use of 4-year weights and updates the weight variable name.
\item If the original weight name is \code{WTDRD1}, it uses the special 4-year dietary weight \code{WTDR4YR}.
\item Updates the weights in the \code{merged_tab}:
\itemize{
\item If only series A and B are present, sets the weight to the new 4-year weight.
\item Otherwise, adjusts weights for series 1 and 2 and averages weights across all series.
}
\item If both series A and B are not present, logs that it is averaging 2-year weights across the surveys and divides the weight by the number of series.
}
}
\examples{
\dontrun{
get_tables_obj <- figure_out_weight(get_tables_obj)
updated_obj <- figure_out_multiyear_weight(get_tables_obj)
}
}
