% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantpe.R
\name{figure_out_weight}
\alias{figure_out_weight}
\title{Identify the appropriate weight column from the exposure and phenotype tables}
\usage{
figure_out_weight(get_tables_obj)
}
\arguments{
\item{get_tables_obj}{A list object produced by the \code{get_tables()} function. Contains the exposure table (e_table), phenotype table (p_table), merged table (merged_tab), and series name.}
}
\value{
The input list object with an additional column "wt" in the merged table (merged_tab) corresponding to the identified weight column.
}
\description{
This function identifies the appropriate weight column from the exposure (e_table) and phenotype (p_table) tables, contained within the list object produced by the \code{get_tables()} function. The weight column is determined based on the following criteria:
\enumerate{
\item If a column named "WTDRD1" exists, it is selected.
\item If not, and a column ending in "2YR" exists, that is selected.
\item If neither of the above conditions are met, the first weight column is selected.
If there is no weight column in either table, a default weight column ("WTMEC2YR") is used.
Once the weight column is identified, a new column "wt" is added to the merged table (merged_tab), which is equivalent to the weight column.
}
}
\examples{
\dontrun{
conn <- dbConnect(...)
get_tables_results <- get_tables("phenotype_var", "exposure_var", "series_name", conn, "phenotype_table", "exposure_table")
weighted_tables <- figure_out_weight(get_tables_results)
}
}
