% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantpe.R
\name{get_expo_pheno_tables}
\alias{get_expo_pheno_tables}
\title{Retrieve and merge demographic, exposure, and phenotype tables from a database}
\usage{
get_expo_pheno_tables(con, pheno_table_name, expo_table_name)
}
\arguments{
\item{con}{The database connection.}

\item{pheno_table_name}{Character. The phenotype table name.}

\item{expo_table_name}{Character. The exposure table name.}
}
\value{
A list containing the merged table, the exposure table, the phenotype table, and the series name.
}
\description{
Retrieve and merge demographic, exposure, and phenotype tables from a database
}
\examples{
\dontrun{
conn <- dbConnect(...)
results <- get_expo_pheno_tables(conn, "phenotype_table", "exposure_table")
}
}
