% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_rsq.R
\name{get_individual_level_table}
\alias{get_individual_level_table}
\title{Retrieve individual level data table based on a set of a variables}
\usage{
get_individual_level_table(summary_stats_con, nhanes_con, pvarname_to_query)
}
\arguments{
\item{summary_stats_con}{A DBI connection object to the summary statistics database.}

\item{nhanes_con}{A DBI connection object to the NHANES database.}

\item{pvarname_to_query}{The variable name to be queried from the database.}
}
\value{
A list containing the final selected variables (\code{evars}), the selected weight to use (\code{weight_to_use}),
and the selected data (\code{selected_data}).
}
\description{
Assumes that findings have already been associated with a phenotype: |> filter(sig_levels == 'Bonf.<0.05')
}
\details{
This function queries individual level data from a provided database connection
based on specific parameters and performs several operations including joining,
filtering, summarizing, and selecting of data, before returning the final processed data.
}
\examples{
\dontrun{
result <- get_individual_level_table(summary_stats_con, nhanes_con, "varname")
}
}
