% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petable.R
\name{get_multiple_x_y_tables}
\alias{get_multiple_x_y_tables}
\title{Get Multiyear X and Y Tables -- possibly delete this, or use sparingly without weighting}
\usage{
get_multiple_x_y_tables(con, table_list_1, table_list_2)
}
\arguments{
\item{con}{A database connection object.}

\item{table_list_1}{A character vector of table names for the first set of tables to be merged.}

\item{table_list_2}{A character vector of table names for the second set of tables to be merged.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{merged_tab}: A data frame of the merged tables.
\item \code{table1}: A data frame of the first set of tables.
\item \code{table2}: A data frame of the second set of tables.
\item \code{series}: A data frame with series information.
}
}
\description{
This function retrieves and merges demographic, table 1, and table 2 data for multiyear NHANES datasets from a specified database connection. It ensures the tables are appropriately joined and returns the merged result along with individual tables and series information.
}
\details{
The function performs the following steps:
\itemize{
\item Retrieves table names from the \code{table_names_epcf} table.
\item Obtains series information based on \code{table_list_1}.
\item Retrieves demographic table names corresponding to the series.
\item Binds demographic tables, \code{table_list_1}, and \code{table_list_2}.
\item Merges the demographic data with the two sets of tables by the \code{SEQN} identifier.
\item Collects the final merged table and returns it along with individual tables and series information.
}
}
\examples{
\dontrun{
con <- DBI::dbConnect(RSQLite::SQLite(), dbname = "nhanes.db")
table_list_1 <- c("table1_1999", "table1_2001")
table_list_2 <- c("table2_1999", "table2_2001")
result <- get_multiyear_x_y_tables(con, table_list_1, table_list_2)
}
}
