% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantpe.R
\name{get_mv_expo_pheno_tables}
\alias{get_mv_expo_pheno_tables}
\title{Retrieve and merge demographic, multi-variable exposure, and phenotype tables from a database}
\usage{
get_mv_expo_pheno_tables(con, pheno_table_name, expo_table_names)
}
\arguments{
\item{con}{The database connection.}

\item{pheno_table_name}{Character. The name of the phenotype table.}

\item{expo_table_names}{Character vector. The names of the exposure tables.}
}
\value{
A list containing the merged table, the final exposure table after joining all exposure tables, the phenotype table, and the series name.
}
\description{
This function retrieves and merges tables from a database based on provided table names. Specifically, it merges demographic, multiple exposure variables and phenotype tables. The function logs the series of the phenotype, the name of the demographics table, the number of rows in each exposure and phenotype table, and finally the number of rows in the merged table.
}
\examples{
\dontrun{
conn <- connect_pewas_data()
results <- get_mv_expo_pheno_tables(conn, "L10AM_C", c("L45VIT_C", "L06COT_C"))
}
}
