% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_rsq.R
\name{get_mv_expo_pheno_tables_for_big_table}
\alias{get_mv_expo_pheno_tables_for_big_table}
\title{Retrieve multivariate exposure and phenotype tables for a large table}
\usage{
get_mv_expo_pheno_tables_for_big_table(con, pheno_table_name, expo_table_names)
}
\arguments{
\item{con}{A DBI connection object to the database.}

\item{pheno_table_name}{The name of the phenotype table.}

\item{expo_table_names}{A vector containing the names of the exposure tables.}
}
\value{
A list containing the final merged table (\code{merged_tab}), the exposure table (\code{e_table}),
the phenotype table (\code{p_table}), the series name (\code{series}), and a tibble containing information
about the table weights (\code{table_weights}).
}
\description{
This function retrieves exposure and phenotype tables based on the provided table names,
performs several operations including joining and filtering, and returns a list containing
the final processed tables and relevant information.
}
\examples{
\dontrun{
result <- get_mv_expo_pheno_tables_for_big_table(con, "pheno_table", c("expo_table1", "expo_table2"))
}
}
