% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantpe.R
\name{get_tables}
\alias{get_tables}
\title{Retrieve and merge tables from a database based on the series name, exposure variable name, and phenotype variable name}
\usage{
get_tables(
  pvarname,
  evarname,
  seriesName,
  con,
  pheno_table_name = NULL,
  expo_table_name = NULL
)
}
\arguments{
\item{pvarname}{Character. The phenotype variable name.}

\item{evarname}{Character. The exposure variable name.}

\item{seriesName}{Character. The series name.}

\item{con}{The database connection.}

\item{pheno_table_name}{Character. The optional phenotype table name. If NULL (default), the function will attempt to find the table based on the \code{pvarname}.}

\item{expo_table_name}{Character. The optional exposure table name. If NULL (default), the function will attempt to find the table based on the \code{evarname}.}
}
\value{
A list containing the merged table, the exposure table, the phenotype table, and the series name.
}
\description{
Retrieve and merge tables from a database based on the series name, exposure variable name, and phenotype variable name
}
\examples{
\dontrun{
conn <- dbConnect(...)
results <- get_tables("phenotype_var", "exposure_var", "series_name", conn, "phenotype_table", "exposure_table")
}
}
