% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petable.R
\name{get_x_y_tables}
\alias{get_x_y_tables}
\title{Retrieve and merge demographic, and any two tables from the nhanes database}
\usage{
get_x_y_tables(con, table_name1, table_name2)
}
\arguments{
\item{con}{The database connection.}

\item{table_name1}{Character. table name 1.}

\item{table_name2}{Character. table name 2}
}
\value{
A list containing the merged table, the table 1, the table 2, and the series name.
}
\description{
Retrieve and merge demographic, and any two tables from the nhanes database
}
\examples{
\dontrun{
conn <- connect_pewas_data()
results <- get_expo_pheno_tables(conn, pheno_table_name="L10AM_C", expo_table_name="L45VIT_C")
}
}
