% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantpe.R
\name{pe}
\alias{pe}
\title{Phenotype and Exposure Association Across Surveys}
\usage{
pe(
  pheno,
  exposure,
  adjustment_variables,
  con,
  series = NULL,
  logxform_p = T,
  logxform_e = T,
  scale_e = T,
  scale_p = F,
  pheno_table_name = NULL,
  expo_table_name = NULL,
  quantile_expo = NULL,
  exposure_levels = NULL
)
}
\arguments{
\item{pheno}{A character string specifying the phenotype variable.}

\item{exposure}{A character string specifying the exposure variable.}

\item{adjustment_variables}{A character vector of adjustment variables that must be present in the demographic table.}

\item{con}{A database connection object.}

\item{series}{A character vector specifying the series to consider (default is NULL).}

\item{logxform_p}{Logical, whether to log-transform the phenotype variable (default is TRUE).}

\item{logxform_e}{Logical, whether to log-transform the exposure variable (default is TRUE).}

\item{scale_e}{Logical, whether to scale the exposure variable (default is TRUE).}

\item{scale_p}{Logical, whether to scale the phenotype variable (default is FALSE).}

\item{pheno_table_name}{Optional character string specifying the phenotype table name (default is NULL).}

\item{expo_table_name}{Optional character string specifying the exposure table name (default is NULL).}

\item{quantile_expo}{Optional vector specifying quantiles for the exposure variable (default is NULL).}

\item{exposure_levels}{Optional vector specifying levels for the exposure variable (default is NULL).}
}
\value{
A list containing the following elements:
\itemize{
\item \code{dsn}: The survey design object.
\item \code{unweighted_n}: The number of unweighted observations.
\item \code{phenotype}: The phenotype variable.
\item \code{exposure}: The exposure variable.
\item \code{series}: The series information.
\item \code{unadjusted_model}: The unadjusted model.
\item \code{adjusted_model}: The adjusted model.
\item \code{base_mod}: The base model adjusted for the adjustment variables.
\item \code{demo_breakdown}: A table with demographic breakdown of the survey design.
}
}
\description{
This function performs analysis of phenotype and exposure with adjustment variables. It handles the retrieval of necessary data, weighting, and running of models, and provides a demographic breakdown.
}
\details{
The function performs the following steps:
\itemize{
\item If a single series is provided, calls \code{pe_by_survey_series} for analysis.
\item Retrieves table names for the phenotype and exposure variables.
\item Ensures the phenotype and exposure variables are collected in the same survey.
\item Gets table names for each series and binds the tables.
\item Handles weights by calling \code{figure_out_multiyear_weight}.
\item Transforms and names the phenotype and exposure variables if specified.
\item Filters data and creates a survey design object.
\item Performs demographic breakdown of the survey design.
\item Runs unadjusted, adjusted, and base models for the analysis.
}
}
\examples{
\dontrun{
con <- DBI::dbConnect(RSQLite::SQLite(), dbname = "nhanes.db")
adjustment_vars <- c("RIDAGEYR", "RIAGENDR")
result <- pe(pheno = "BMXBMI", exposure = "LBXCOT", adjustment_variables = adjustment_vars, con = con)
}
}
