% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantpe.R
\name{pe_by_survey_series}
\alias{pe_by_survey_series}
\title{P-E association by a single survey}
\usage{
pe_by_survey_series(
  pheno,
  exposure,
  adjustment_variables,
  series,
  con,
  logxform_p = T,
  logxform_e = T,
  scale_e = T,
  scale_p = F,
  pheno_table_name = NULL,
  expo_table_name = NULL,
  quantile_expo = NULL,
  exposure_levels = NULL
)
}
\arguments{
\item{pheno}{A variable related to phenotype}

\item{exposure}{A variable related to exposure}

\item{adjustment_variables}{A vector of variables to adjust for in the models. They have to be in the demo table.}

\item{series}{A variable related to series}

\item{con}{Connection object for the NHANES database}

\item{logxform_p}{Logical, if TRUE, a log transformation is applied to pheno variable. Default is TRUE.}

\item{logxform_e}{Logical, if TRUE, a log transformation is applied to exposure variable. Default is TRUE.}

\item{scale_e}{Logical, if TRUE, exposure variable is scaled. Default is TRUE.}

\item{scale_p}{Logical, if TRUE, pheno variable is scaled. Default is FALSE.}

\item{pheno_table_name}{Optional, name of the phenotype table}

\item{expo_table_name}{Optional, name of the exposure table}

\item{quantile_expo}{Optional, if specified, exposure variable will be cut by these quantiles}

\item{exposure_levels}{Optional, if specified, exposure variable will be treated as a categorical variable with these levels}
}
\value{
A list containing: the svydesign object, unweighted number of observations, phenotype, exposure, series, unadjusted model, adjusted model, base model, and demographic breakdown.
}
\description{
Execute an association between a P and E variable
This function performs a series of operations including fetching tables,
applying weight adjustments, performing log transformations, creating survey designs,
running models, and computing demographic breakdowns.
}
\examples{
\dontrun{
connection <- connect_pewas_data()
adjustments <- adjustment_variables()
pe_result <- pe(pheno="BMXBMI", exposure="LBXGTC", adjustment_variables=adjustments, series="C",
con=connection, logxform_p=T, logxform_e=T, scale_e=T, scale_p=F)
pe_result$adjusted_model$tidied
}

}
