% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantpe.R
\name{pe_by_table}
\alias{pe_by_table}
\title{PE by Table}
\usage{
pe_by_table(
  tab_obj,
  pvar,
  evar,
  adjustment_variables,
  logxform_p = T,
  logxform_e = T,
  scale_e = T,
  scale_p = F,
  quantile_expo = NULL,
  exposure_levels = NULL
)
}
\arguments{
\item{tab_obj}{A pre-processed table object including both phenotype and exposure information}

\item{pvar}{A variable related to phenotype}

\item{evar}{A variable related to exposure}

\item{adjustment_variables}{A vector of variables to adjust for in the models}

\item{logxform_p}{Logical, if TRUE, a log transformation is applied to phenotype variable. Default is TRUE.}

\item{logxform_e}{Logical, if TRUE, a log transformation is applied to exposure variable. Default is TRUE.}

\item{scale_e}{Logical, if TRUE, exposure variable is scaled. Default is TRUE.}

\item{scale_p}{Logical, if TRUE, phenotype variable is scaled. Default is FALSE.}

\item{quantile_expo}{Optional, if specified, exposure variable will be cut by these quantiles}

\item{exposure_levels}{Optional, if specified, exposure variable will be treated as a categorical variable with these levels}
}
\value{
A list containing: the svydesign object, log transformation status for phenotype and exposure, scaling status for phenotype and exposure, unweighted number of observations, phenotype, series, exposure, unadjusted model, adjusted model, base model, and demographic breakdown.
}
\description{
This function operates similarly to the 'pe' function, but takes a pre-processed table object as input instead of separate phenotype and exposure tables. It is designed to help users scale up the pe associations by executing all possible associations between the P and E tables
}
\examples{
\dontrun{
pe_result <- pe_by_table(tab_obj=table_object, pvar="BMI", evar="smoke", adjustment_variables=c("age", "sex"), logxform_p=T, logxform_e=T, scale_e=T, scale_p=F, quantile_expo=NULL, exposure_levels=NULL)
}

}
