% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_rsq.R
\name{pme}
\alias{pme}
\title{PME - multivariate regression: P ~ E1 + E2 + ...}
\usage{
pme(tab_obj, pvar, evars, adjustment_variables, logxform_p = T, scale_p = F)
}
\arguments{
\item{tab_obj}{A pre-processed table object including both phenotype and exposure information.}

\item{pvar}{A variable related to phenotype.}

\item{evars}{A vector of variables related to multiple exposures.}

\item{adjustment_variables}{A vector of variables to adjust for in the models.}

\item{logxform_p}{Logical, if TRUE, a log transformation is applied to phenotype variable. Default is TRUE.}

\item{scale_p}{Logical, if TRUE, phenotype variable is scaled. Default is FALSE.}
}
\value{
A list containing: the svydesign object, log transformation status for phenotype, scaling status for phenotype, unweighted number of observations, phenotype, series, exposures, unadjusted model, adjusted model, and base model.
}
\description{
This function performs a multivariate exposure (multiple exposures) regression analysis on a pre-processed table object. The phenotype can be log-transformed and scaled if desired, and the model can be adjusted for specified variables.
}
\examples{
\dontrun{
table_object <- get_mv_expo_pheno_tables(conn, "L10AM_C", c("L45VIT_C", "L06COT_C"))
table_object <- figure_out_weight(table_object)
pme_result <- pme(tab_obj=table_object, pvar="LBXGLU", evars=c("LBXCOT", "LBXGTC"),
adjustment_variables=adjustment_variables(), logxform_p=T, scale_p=F)
}

}
