% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpe.R
\name{run_logistic_model}
\alias{run_logistic_model}
\title{Run Logistic Model on Survey Data}
\usage{
run_logistic_model(
  formu,
  dsn,
  scale_expo = T,
  quantile_expo = NULL,
  expo_levels = NULL,
  save_svymodel = F
)
}
\arguments{
\item{formu}{A formula object specifying the logistic regression model to fit.}

\item{dsn}{A survey design object created using functions like \code{svydesign} from the \strong{survey} package.}

\item{scale_expo}{Logical; if \code{TRUE}, the exposure variable will be scaled (z-score transformation). Default is \code{TRUE}.}

\item{quantile_expo}{Numeric vector of quantiles (e.g., \code{c(0, 0.25, 0.5, 0.75, 1)}) to discretize the exposure variable into quantile bins. Default is \code{NULL}.}

\item{expo_levels}{Character vector of specific levels to treat the exposure as a categorical variable. Default is \code{NULL}.}

\item{save_svymodel}{Logical; if \code{TRUE}, the full survey model object is included in the output. Default is \code{FALSE}.}
}
\value{
A list with the following elements:
\item{glanced}{A tibble with model summary statistics produced by \code{broom::glance}.}
\item{tidied}{A tibble with coefficient estimates and related statistics produced by \code{broom::tidy}.}
\item{scale_expo}{Logical, indicating whether the exposure variable was scaled.}
\item{q_cut_points}{Numeric vector of cut points used if the exposure was discretized into quantiles.}
\item{quantiles}{The input vector of quantiles used for discretization (if applicable).}
\item{expo_levels}{The input levels of the categorical exposure variable (if applicable).}
\item{model}{The full survey model object, if \code{save_svymodel = TRUE}; otherwise, \code{NA}.}
}
\description{
This function fits a logistic regression model using survey data, allowing for the scaling, quantile transformation, or categorical recoding of an exposure variable.
}
\details{
This function is designed to run logistic regression models on survey data, providing flexibility for handling exposure variables. It supports:
\itemize{
\item Scaling the exposure variable (z-score transformation).
\item Discretizing the exposure into quantile bins.
\item Treating the exposure as a categorical variable with specified levels.
}

The function utilizes the \strong{survey} package for modeling, with \code{quasibinomial} as the default family for logistic regression.
}
\examples{
\dontrun{
library(survey)
library(broom)

# Example survey data
data(api)
dstrat <- svydesign(id = ~1, strata = ~stype, weights = ~pw, data = apistrat, fpc = ~fpc)

# Run logistic model with scaling
result <- run_logistic_model(api00 ~ ell + meals, dstrat, scale_expo = TRUE)

# Access results
result$glanced
result$tidied
}

}
