% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantpe.R
\name{svyrsquared}
\alias{svyrsquared}
\title{Calculate R-squared and adjusted R-squared for a survey-weighted linear model}
\usage{
svyrsquared(analysisObj)
}
\arguments{
\item{analysisObj}{A list containing the components necessary for the calculation.
It should have the following components:
\itemize{
\item survey.design: A survey design object.
\item residuals: A numeric vector of residuals from the model.
\item y: A numeric vector of the response variable.
}}
}
\value{
A list with two components:
\itemize{
\item rsq: R-squared of the linear model.
\item adj.r2: Adjusted R-squared of the linear model.
}
}
\description{
This function calculates the R-squared and adjusted R-squared values for a
linear model based on an analysis object which containssave   a survey design,
residuals, and a response variable. The calculation is conducted using
survey-weighted means, totals, and residuals.
}
\examples{
\dontrun{
analysisObj <- list(survey.design = dsn, residuals = resid, y = y)
results <- svyrsquared(analysisObj)
}
}
