% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantpe.R
\name{xy_by_table_flex_adjust}
\alias{xy_by_table_flex_adjust}
\title{XY by Table Flexible Adjustments}
\usage{
xy_by_table_flex_adjust(
  tab_obj,
  yvar,
  xvar,
  adjustment_variables,
  logxform_x = T,
  logxform_y = T,
  scale_x = T,
  scale_y = T,
  quantile_expo = NULL,
  exposure_levels = NULL
)
}
\arguments{
\item{tab_obj}{A pre-processed table object including both phenotype and exposure information}

\item{yvar}{A variable related to phenotype}

\item{xvar}{A variable related to exposure}

\item{adjustment_variables}{A tibble of variables to adjust for in the models (to consider multiple scenarios)}

\item{logxform_x}{Logical, if TRUE, a log transformation is applied to exposure variable. Default is TRUE.}

\item{logxform_y}{Logical, if TRUE, a log transformation is applied to phenotype variable. Default is TRUE.}

\item{scale_y}{Logical, if TRUE, phenotype variable is scaled. Default is FALSE.}
}
\value{
A list containing: the svydesign object, log transformation status for phenotype and exposure, scaling status for phenotype and exposure, unweighted number of observations, phenotype, series, exposure, models (per adjustment) and demographic breakdown.
}
\description{
This function operates similarly to the 'pe' function, but takes a pre-processed table object as input instead of separate phenotype and exposure tables. It is designed to help analysts scale up the pe associations by sepearting the "get_tables" procedure from runnin an association.
It also takes a tibble for the adjustment variables to flexibly parameterize adjustments
}
\examples{
\dontrun{
con <- connect_pewas_data()
table_object <- get_tables("LBXGLU", "LBXGTC", "C", con, "L10AM_C", "L45VIT_C")
pe_result <- xy_by_table_flex_adjust(tab_obj=table_object, yvar="LBXGLU", xvar="LBXGTC",
adjustment_variables=c("RIDAGEYR", "RIAGENDR"), # tibble
}

}
