% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantpe.R
\name{run_model}
\alias{run_model}
\title{Run a model with options for scaling and quantile-based categorization of exposure and phenotype variables}
\usage{
run_model(
  formu,
  dsn,
  scale_expo = T,
  scale_pheno = F,
  quantile_expo = NULL,
  expo_levels = NULL,
  save_svymodel = F
)
}
\arguments{
\item{formu}{A formula specifying the model to be run.}

\item{dsn}{A survey design object.}

\item{scale_expo}{Logical, if TRUE, the exposure variable is scaled by its standard deviation.}

\item{scale_pheno}{Logical, if TRUE, the phenotype variable is scaled by its standard deviation.}

\item{quantile_expo}{Numeric vector, quantiles to use for categorizing the exposure variable.
If not NULL, this overwrites the scale_expo argument.}

\item{expo_levels}{A vector of levels to categorize the exposure variable.}

\item{save_dsn}{Logical, if TRUE, the desgin object for each model will be saved
If not NULL, this overwrites the quantile_expo and scale_expo arguments.}
}
\value{
A list containing:
\itemize{
\item model: The model object (if the parameter is set to T)
\item glanced: A one-row data frame of model-level statistics from glance().
\item tidied: A data frame of term-level statistics from tidy().
\item r2: R-squared and adjusted R-squared from svyrsquared().
\item scale_pheno: Whether the phenotype variable was scaled.
\item scale_expo: Whether the exposure variable was scaled.
\item q_cut_points: Cut points used for quantile-based categorization of the exposure variable.
\item quantiles: The input quantiles for categorizing the exposure variable.
\item expo_levels: The input levels for categorizing the exposure variable.
}
}
\description{
This function runs a survey-weighted generalized linear model based on the given formula and design,
with options for scaling the exposure and phenotype variables and for categorizing the exposure variable
based on quantiles or given levels.
}
\examples{
\dontrun{
result <- run_model(formu, dsn, scale_expo=T, scale_pheno=F, quantile_expo=c(0.25, 0.5, 0.75))
}
}
